+++ ThatDayThatTime/TimeDiaryWidget/Entry.swift
@@ -1,0 +1,13 @@
+//
+//  Entry.swift
+//  ThatDayThatTime
+//
+//  Created by J_Min on 7/1/24.
+//
+
+import WidgetKit
+
+struct DiaryEntry: TimelineEntry {
+    let date: Date
+    let diary: DiaryEntity
+}
